RTL Support - swap elements
All my suggested changes are designated by comments. Please, feel free to accept these changes as they are or modify them in any way you like. Moreover, please feel free to remove all of my comments.

I have tested the changes on the js script, which is in the dist folder, which is installed by npm. I have NOT tested the code in this js in the src folder! You might find a silly typo/bug.